# docs/swagger.yaml
basePath: /
definitions:
  github_com_fluentfox_api_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_fluentfox_api_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorBody'
    type: object
...
  internal_srs.CardResponse:
    properties:
      card_face:
        type: string
      card_type:
        type: string
      content_id:
        type: string
      due:
        type: string
      id:
        type: string
      is_new:
        type: boolean
      lapses:
        type: integer
    type: object
  internal_srs.DueCountResponse:
    properties:
      new:
        type: integer
      review:
        type: integer
      total:
        type: integer
    type: object
  internal_srs.ReviewResultResponse:
    properties:
      card_id:
        type: string
      is_leech:
        type: boolean
      next_due:
        type: string
      scheduled_days:
        type: integer
      xp_earned:
        type: integer
    type: object
  internal_srs.SubmitReviewRequest:
    properties:
      card_id:
        type: string
      rating:
        maximum: 4
        minimum: 1
        type: integer
    required:
    - card_id
    - rating
    type: object
...
  internal_user.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
  internal_user.ProfileResponse:
    properties:
      bio:
        type: string
      display_name:
        type: string
      jlpt_goal:
        type: string
      native_language:
        type: string
    type: object
  internal_user.SettingsResponse:
    properties:
      daily_xp_goal:
        type: integer
      email_reminders:
        type: boolean
      reminder_time:
        type: string
      srs_cards_per_session:
        type: integer
    type: object
  internal_user.UpdateProfileRequest:
    properties:
      bio:
        maxLength: 300
        type: string
      display_name:
        maxLength: 50
        type: string
      jlpt_goal:
        enum:
        - N5
        - N4
        - N3
        - N2
        - N1
        type: string
      native_language:
        maxLength: 50
        type: string
    type: object
  internal_user.UpdateSettingsRequest:
    properties:
      daily_xp_goal:
        maximum: 500
        minimum: 10
        type: integer
      email_reminders:
        type: boolean
      reminder_time:
        type: string
      srs_cards_per_session:
        maximum: 100
        minimum: 5
        type: integer
    type: object
  internal_user.UserResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  internal_xp.LeaderboardEntry:
    properties:
      level:
        type: integer
      rank:
        type: integer
      total_xp:
        type: integer
      user_id:
        type: string
      username:
        type: string
    type: object
  internal_xp.XPResponse:
    properties:
      level:
        type: integer
      progress_percent:
        type: number
      total_xp:
        type: integer
      xp_to_next_level:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Japanese language learning API — chapters, SRS flashcard reviews, quizzes, XP levelling, streaks, and in-app shop.
    All successful responses are wrapped in {"data": ...}. Paginated responses also include {"meta": {...}}.
    All error responses follow {"error": {"code": "...", "message": "..."}}.
  title: FluentFox API
  version: "1.0"
paths:
...
  /srs/due:
    get:
      description: Returns cards that are due for review today, ordered by due date.
        Limit defaults to the user's deck setting.
      parameters:
      - description: Maximum number of cards to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_srs.CardResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get due SRS cards
      tags:
      - srs
  /srs/due/count:
    get:
      description: Returns counts of new cards and review cards due today.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_srs.DueCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get due card count
      tags:
      - srs
  /srs/review:
    post:
      consumes:
      - application/json
      description: Records a review rating (1=Again, 2=Hard, 3=Good, 4=Easy) and reschedules
        the card using the FSRS algorithm. Awards XP.
      parameters:
      - description: Card ID and rating
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_srs.SubmitReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_srs.ReviewResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a card review
      tags:
      - srs
...
  /users/me:
    get:
      description: Returns the authenticated user's account details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /users/me/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password after verifying the current
        one.
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Current password incorrect
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
  /users/me/profile:
    get:
      description: Returns the authenticated user's public profile (display name,
        bio, JLPT goal).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates display name, bio, native language, and/or JLPT goal.
      parameters:
      - description: Profile fields to update (all optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /users/me/settings:
    get:
      description: Returns the authenticated user's app preferences (daily XP goal,
        SRS card count, reminders).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.SettingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user settings
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates one or more app preferences for the authenticated user.
      parameters:
      - description: Settings fields to update (all optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - users
  /xp:
    get:
      description: Returns the user's total XP, current level, XP to next level, and
        progress percentage.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_xp.XPResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get XP and level
      tags:
      - xp
  /xp/leaderboard:
    get:
      description: Returns the global XP leaderboard ranked by total XP.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_xp.LeaderboardEntry'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get XP leaderboard
      tags:
      - xp
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_access_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"